Monolith Service is ready...

+++ OnlineShop/src/main/resources/db/changelog/001_add_all_entity.sql
@@ -1,6 +1,6 @@
 
 CREATE TABLE t_user (
-                         id serial PRIMARY KEY,
+                         id BIGSERIAL PRIMARY KEY,
                          username VARCHAR(255) NOT NULL UNIQUE,
                          first_name VARCHAR(255) NOT NULL,
                          last_name VARCHAR(255) NOT NULL,
@@ -9,8 +9,8 @@ CREATE TABLE t_user (
 );
 
 CREATE TABLE t_user_info (
-                         id serial PRIMARY KEY,
-                         user_id INT,
+                         id BIGSERIAL PRIMARY KEY,
+                         user_id BIGINT,
                          email VARCHAR(255) NOT NULL UNIQUE,
                          address_line VARCHAR(255),
                          city VARCHAR(255) NOT NULL,
@@ -21,16 +21,16 @@ CREATE TABLE t_user_info (
 );
 
 CREATE TABLE t_role (
-                         id serial PRIMARY KEY,
+                         id BIGSERIAL PRIMARY KEY,
                          name VARCHAR(255) NOT NULL UNIQUE,
                          description VARCHAR(255)
 );
 
 CREATE TABLE t_user_auth (
-                             id serial PRIMARY KEY,
+                             id BIGSERIAL PRIMARY KEY,
                              password TEXT,
-                             user_id INT,
-                             role_id INT,
+                             user_id BIGINT ,
+                             role_id BIGINT ,
                              token TEXT,
                              expired_date TIMESTAMP NOT NULL,
                              is_expired BOOLEAN NOT NULL,
@@ -39,8 +39,8 @@ CREATE TABLE t_user_auth (
 );
 
 CREATE TABLE t_user_payment (
-                                id serial PRIMARY KEY,
-                                user_id INT NOT NULL,
+                                id BIGSERIAL PRIMARY KEY,
+                                user_id BIGINT  NOT NULL,
                                 payment_type VARCHAR(255) NOT NULL,
                                 provider VARCHAR(255) NOT NULL,
                                 account_no VARCHAR(255) NOT NULL,
@@ -49,8 +49,8 @@ CREATE TABLE t_user_payment (
 );
 
 CREATE TABLE t_shopping_session (
-                                id serial PRIMARY KEY,
-                                user_id INT NOT NULL,
+                                id BIGSERIAL PRIMARY KEY,
+                                user_id BIGINT  NOT NULL,
                                 total DECIMAL DEFAULT 0,
                                 created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                                 modified_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -58,7 +58,7 @@ CREATE TABLE t_shopping_session (
 );
 
 CREATE TABLE t_product_category (
-                                    id serial PRIMARY KEY,
+                                    id BIGSERIAL PRIMARY KEY,
                                     name VARCHAR(255) NOT NULL,
                                     description TEXT NOT NULL,
                                     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -68,7 +68,7 @@ CREATE TABLE t_product_category (
 
 
 CREATE TABLE t_product_inventory (
-                                    id serial PRIMARY KEY,
+                                    id BIGSERIAL PRIMARY KEY,
                                     quantity INT NOT NULL,
                                     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                                     modified_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -76,7 +76,7 @@ CREATE TABLE t_product_inventory (
 );
 
 CREATE TABLE t_discount (
-                                     id serial PRIMARY KEY,
+                                     id BIGSERIAL PRIMARY KEY,
                                      name VARCHAR(255) NOT NULL,
                                      description TEXT,
                                      discount_percent DECIMAL NOT NULL,
@@ -89,14 +89,14 @@ CREATE TABLE t_discount (
 
 
 CREATE TABLE t_product (
-                                id serial PRIMARY KEY,
+                                id BIGSERIAL PRIMARY KEY,
                                 name VARCHAR(255) NOT NULL,
                                 description TEXT,
                                 SKU VARCHAR(255),
-                                category_id INT,
-                                inventory_id INT,
+                                category_id BIGINT ,
+                                inventory_id BIGINT ,
                                 price DECIMAL,
-                                discount_id INT,
+                                discount_id BIGINT ,
                                 created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                                 modified_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                                 deleted_at TIMESTAMP DEFAULT NULL,
@@ -106,15 +106,15 @@ CREATE TABLE t_product (
 );
 
 CREATE TABLE t_payment_status (
-                                id serial PRIMARY KEY,
+                                id BIGSERIAL PRIMARY KEY,
                                 name VARCHAR(255) NOT NULL UNIQUE
 );
 
 
 CREATE TABLE t_cart_item (
-                             id serial PRIMARY KEY,
-                             session_id INT,
-                             product_id INT,
+                             id BIGSERIAL PRIMARY KEY,
+                             session_id BIGINT ,
+                             product_id BIGINT ,
                              quantity INT,
                              created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                              modified_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -124,8 +124,8 @@ CREATE TABLE t_cart_item (
 
 
 CREATE TABLE t_order_details (
-                                 id serial PRIMARY KEY,
-                                 user_id INT,
+                                 id BIGSERIAL PRIMARY KEY,
+                                 user_id BIGINT ,
                                  total DECIMAL,
                                  created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                                  modified_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -133,9 +133,9 @@ CREATE TABLE t_order_details (
 );
 
 CREATE TABLE t_order_items (
-                             id serial PRIMARY KEY,
-                             order_id INT,
-                             product_id INT,
+                             id BIGSERIAL PRIMARY KEY,
+                             order_id BIGINT ,
+                             product_id BIGINT ,
                              quantity INT,
                              created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                              modified_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
@@ -144,13 +144,14 @@ CREATE TABLE t_order_items (
 );
 
 CREATE TABLE t_payment_details (
-                                   id serial PRIMARY KEY,
-                                   order_id INT,
+                                   id BIGSERIAL PRIMARY KEY,
+                                   order_id BIGINT ,
                                    amount INT,
                                    provider VARCHAR(255),
-                                   status_id INT,
+                                   status_id BIGINT ,
                                    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                                    modified_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
                                    FOREIGN KEY (status_id) REFERENCES t_payment_status(id) ON DELETE SET NULL,
                                    FOREIGN KEY (order_id) REFERENCES t_order_details(id) ON DELETE CASCADE
 );
+
